@marshallpierce 2018-11-20
Thanks, SIMD has been on my todo list for a long time. Why have merely fast base64 when it could be completely unnecessarily fast?

Can you split out rustfmt/clippy/criterion into its own PR to make it easier to see which are the SIMD changes?

I won't have time for a detailed review of the SIMD work for a while, but just as food for thought -- a design I explored in https://bitbucket.org/marshallpierce/stream-vbyte-rust/src that worked out pretty well was to have a trait that represented the core encoding/decoding logic with implementations for scalar and all the relevant flavors of SIMD. That helped keep it (arguably?) more readable and maintainable. I'd be inclined to follow that to make conceptual room for NEON implementations, etc, but I haven't taken a close look at what you've got here yet.