@ggriffiniii 2018-11-21
Yes, I can split out the rustfmt/clippy/criterion changes into a separate PR. Would you like those as 3 PR's or just 1?

Regarding the SIMD implementation. I had considered a trait and it would be a straightforward change if you prefer. In this case I didn't think it was worth it since it's just a single function (bulk_encode_slice) and putting the function within a module seemed like it accomplished much of the same goal as a trait would, but I can see how a trait could potentially evolve more gracefully when adding decode, other architectures, etc.

I noticed that the continuous build failed trying to build against rust 1.23. Is that the minimum rust version you're targeting? If so, this may not be possible yet. This change would require 1.27 or later if the simd functionality is enabled.