@ggriffiniii 2018-11-21
Also I'm sure you realize, but GitHub will show diffs for individual commits if you click on them. I'm fine splitting the change into multiple pull requests still. Just felt obligated to point that out.