@marshallpierce 2018-11-21
Just one PR for the precursor commits is fine -- that one I should be able to get merged quickly. Might as well have that much done and easily available for other work.

Let's leave the code structure as is for now. It'll be easy enough to switch to traits if need be.

SIMD is definitely worth bumping the minimum rust version for.