@marshallpierce 2018-11-21
Also worth looking into from my stash o' bookmarks, if you hadn't seen them already:

https://github.com/lemire/fastbase64
https://arxiv.org/pdf/1704.00605.pdf
http://richardstartin.uk/base64-encoding/ for more context on the above
http://0x80.pl/notesen/2016-04-03-avx512-base64.html for AVX512