@ggriffiniii 2019-07-28
I'm going to close this pull request as it's probably suffered bitrot over the months. If you're still interested in implementing this there is an improved version in my base64 library `radix64` https://github.com/ggriffiniii/radix64 

Feel free to take any bits you like.